requirements.txt
selenium~=4.17.2
# webdriver_manager~=4.0.1
python-dotenv~=1.0.1
python-telegram-bot~=20.8
stackapi>=0.3.0
requests>=2.0.0
requests_oauthlib>=1.3.1